hide v1 better
